Bug 1743647 - Allow VA-API only in the RDD process. r?alwu
The content process sandbox won't necessarily work with VA-API drivers
(e.g., Intel if it uses SysV IPC), and in general we'd like to remove
direct access to the window system / GPU / etc.  Therefore, this patch
restricts VA-API use to the RDD process, where it is intended to work
(with sandboxing enabled).

Differential Revision: https://phabricator.services.mozilla.com/D133716