Bug 1771382 - Adjust the Linux RDD sandbox to handle the nvidia driver being loaded but not used. r?gcp
On multi-GPU systems, even though the GPU we're going to use for
accelerated video decoding is driven by Mesa, sometimes the nvidia
proprietary driver can be loaded and attempt to probe devices.  This
patch attempts to make the sandbox policy quietly return errors for
those syscalls, instead of treating them as unexpected (and crashing on
Nightly).

Differential Revision: https://phabricator.services.mozilla.com/D149652