Bug 1768809 - Fix library preloading for the OpenH264 plugin.
The workaround added in bug 1725828 was intended to be applied to any
plugin that may have been linked against a different version of glibc
than the browser; i.e., everything except clearkey, which is built and
shipped along with the browser.  Unfortunately, the change was made in a
function used only for the Widevine CDM, so we're now having the same
problem with OpenH264.  This patch corrects that oversight and preloads
the (potentially) needed libraries for every applicable plugin.

Differential Revision: https://phabricator.services.mozilla.com/D149446