Bug 1793523 - Remove dead code from DidProcessCrash, leaving a Unix-only IsProcessDead.
Bug 1277705 removed most of the callers of the IPC helper function
DidProcessCrash.  Currently the only use is on Unix (so the Windows
implementation is dead) and only to check if the process exited; the
return value (i.e., did the process crash) is ignored.

To simplify things for the next patch, this patch removes the unused
functionality and leaves a Unix-only IsProcessDead (the name of the
wrapper that was previously the only call site).

No functional change is intended.