Bug 1796758 - Don't try to use `waitid` on OpenBSD. r?nika
OpenBSD, uniquely (as far as I know) of our Unix targets, doesn't
support the function `waitid`, which means that the fix for bug 1793523
broke the build.  However, OpenBSD isn't affected by bug 1793523 because
the crash reporter isn't supported, so it can continue to use `waitpid`.

This patch restores the `waitpid` version of `IsProcessDead` under an
ifdef.  I've tested it locally on Linux (by changing the ifdef), and as
expected it does exhibit bug 1793523 but otherwise seems to work, so it
should work on OpenBSD.

Differential Revision: https://phabricator.services.mozilla.com/D160113