Bug 1784517 - Explicitly skip sandboxing tests on unsandboxed build types. r?glandium
There are tools which consume `moz.build` files by reading every one in
the tree, rather than traversing `DIRS` for a specific build type (see
bug 1667271 comment #89 for background); as a result, we can end up with
CI test jobs that try to run sandboxing tests on build types like Linux
ASan where `security/sandbox` isn't built, and fail.

This patch applies a suggested workaround: add an otherwise redundant
`skip-if` declaration to the test manifests to skip them on the platforms
where they're not part of the build.

Note that sandboxing is disabled in the presence of Linux ASan or TSan
by logic in `toolkit/moz.configure`, but for code coverage builds it's
done via the CI mozconfig files adding `--disable-sandbox`.