Bug 1820542 - Use gdk_pointer_grab instead of gdk_device_grab to avoid bug with nested menus. r?emilio
This is the same issue as bug 1750721, where we react to the pointer
position relative to the wrong window when the pointer is grabbed for a
popup menu.  It was reintroduced in the cluster of bugs that removed and
partly readded our use of pointer grabbing (bug 1807482 in particular).

GTK 3.0 deprecated gdk_pointer_grab in favor of gdk_device_grab, and
their implementations look very similar (as does a trace of the calls
to Grab/Ungrab functions in Xlib), but nonetheless the older function
doesn't exhibit this bug.

Differential Revision: https://phabricator.services.mozilla.com/D176616