Bug 1867673 - Gracefully handle statx when called in sandboxed GMP processes. r?gcp
Testing note: this bug seems to break all GMP plugins on 32-bit x86,
including clearkey, so the regular EME tests would provide coverage, but
we're not testing that platform at all in CI anymore, let alone with a
new enough glibc.  And testing this directly is difficult because our
sandbox tests run with crash-on-error mode turned off, and in that case
it just fails with ENOSYS both before and after.

Differential Revision: https://phabricator.services.mozilla.com/D195194