Bug 1878213 - Add valgrind suppressions for intentional leaks in the forkserver and related code. r?jseward
The fork server replaces the process's `argv` and `environ` with
heap-allocated strings received over IPC; if those strings are later
unlinked from those vectors, they will be seen as leaked.  Also, we
have an implementation of `setproctitle` (a BSD feature that has to be
emulated somewhat hackily on Linux) which copies the process's original
`argv` and `environ` strings into the heap so the original space can be
reused; this can also cause (finite) leaks when those strings are later
discarded.

Given that we already tell Valgrind to ignore leaks caused by use of
`setenv` and similar, this patch likewise suppresses those leaks.

Differential Revision: https://phabricator.services.mozilla.com/D201190