Bug 1881386 - Check for zombie processes in ProcessWatcher on Linux.
FIXME do not land, needs ifdef XP_LINUX and more info on CI zombiecheck

This patch checks whether child processes are in the zombie state
(`exit`ed but not `wait`ed), by parsing `/proc/{pid}/stat`, and treats
them as dead in that case.

Child processes can end up stuck in zombie state if the fork server
exits when child processes are still running, which causes them to be
reparented to pid 1, and pid 1 isn't acting as `init`, which can happen
in container environments like Docker depending on configuration.  In
particular, this is currently the case in the containers used by Mozilla
CI to run tests.  Without this patch, if the fork server is enabled, we
wait forever for the process to exit and then (in the Mozilla CI case)
some other timeout fires and causes the test run to fail.

Longer-term I plan to have fork server wait for child processes, which
means that it will have to outlive them (assuming it doesn't crash), and
that will take care of this.