Bug 1870499 - Crash instead of looping infinitely in `HttpChannelChild::Release`. r?#necko-reviewers
When the last reference to `HttpChannelChild` is dropped, under some
circumstances it can un-drop itself to make sure that cleanup happens
correctly, by posting a runnable to the main thread to invoke a callback.
Unfortunately, under some circumstances this doesn't finalize the object,
causing an infinite loop of runnable dispatch, which causes a shutdown hang.

In particular, on `NS_FREE_PERMANENT_DATA` builds (where we always wait
for child processes to shut down fully — including debug, ASan, and
ccov, but not release builds), the timeout in `ProcessWatcher` obtains
a crash report, which in this case is often in some random code related
to event loops or runnables.  It may be possible for this loop to happen
at times other than shutdown, but I suspect that the underlying cause is
something shutdown releated.

Instead, this patch checks for the same `HttpChannelChild` object
repeatedly looping through that case in `Release`, and crashes in one
convenient place (but only on `NS_FREE_PERMANENT_DATA` builds).  This
doesn't fix the cause of this problem, but at least it will make the
crashes easier to assign to the correct bug.

Differential Revision: https://phabricator.services.mozilla.com/D202780