Bug 1752638 - WORK IN PROGRESS - MiniTransceiver/protocol changes
This patch has my half-finished work making some changes to how fork
server communications works.  Before it used a stream socket and always
had matched pairs of synchronous request/reply with no interruptions.
This changes the main protocol to use a sequential-packet socket with
asynchronous messages in (eventually) both directions; ones that need
a reply (currently the fork request, later also the wait request) will
attach a socket for that, like what we do in the Linux sandbox and crash
reporter.

That also allows requests to be sent from any thread in the main process
without locking, although that will probably need threadsafe refcouting
for safety.  The attached sockets can be stream sockets, e.g. so that
the arg/env strings could exceed the size of the socket buffer if
needed.

What is *not* done here is to actually implement the new messages.
Remoting `waitid` should be relatively straightforward, along similar
lines to how the fork request works after this patch.  Sending `SIGCHLD`
back is more interesting: it's possible to `sendmsg` directly from the
signal handler, but the message would have to be pre-serialized, because
`malloc` is not safe there; it's also possible to use the pipe-to-self
idiom.  In the main process, connecting those messages to the I/O thread
event loop will also need some work.

There is also some subtlety about child processes outliving the fork
server -- currently we intentionally shut it down while child processes
are still running, and it will always be possible for it to crash, so
we'll need some kind of best-effort fallback there.

There are also some FIXME comments in this patch.

And, integration with `ProcessWatcher` should be aware of D141309.

Differential Revision: https://phabricator.services.mozilla.com/D212897