Bug 1896398 - Try to increase fd limit to 64k on (non-Apple) Unix. r?glandium
We're still constrained by any hard limits applied by the distro, but
empirically this is 1024k on Debian/Ubuntu and 512k on Fedora.  In the
past Fedora's limit was only 4k (see bug 1401776 comment #7); both
that value and the current 512k seem to come from systemd's defaults,
and could be changed via systemd config or be overridden via the file
`/etc/security/limits.conf`.

64k is probably enough given that a lot of fd usage is shared memory,
and we're usually limited to 64k mappings due to the vma limit (`sysctl
vm.max_map_count`).

This patch does not change the limit for macOS, because we mostly use
Mach shared memory which doesn't use file descriptors.

Differential Revision: https://phabricator.services.mozilla.com/D212631