Bug 1944914 - Fix ifdef typo (missing E in `SOCK_CLOEXEC`). r?#ipc-reviewers
This typo was missed during review, and then I happened to notice it
myself before landing, and… somehow forgot to fix it.  The impact is
that we won't use `SOCK_CLOEXEC` when creating the socketpair, but we
also won't set `FD_CLOEXEC` non-atomically either, because the other
ifdefs *are* spelled correctly.

It would be nice to have a regression test for this, given how easy it
was for several people to repeatedly miss it, but there isn't a good way
to do that because this is an internal implementation detail.

Differential Revision: https://phabricator.services.mozilla.com/D236307